refactor(core): widen TokenAmount from u64 to u128
Aligns the domain token type with the on-chain uint256 range
(ADR-0001 open question, now resolved). Removes the silent u64 cap that
diverged the off-chain ledger from the contracts.

The domain TokenAmount is now u128 (safe arithmetic, uint256-aligned headroom).
The storage layer and API DTOs remain u64 (practical wire format — real amounts
fit u64) and cast at the boundary. MIN_STAKE stays u64 (tiny constant), cast at
the TokenAmount boundary.

Updated the 4 u64-boundary tests (checked_add/deposit/mint overflow,
total_value saturation) to u128::MAX to preserve their intent.

Tests: 2832 pass (workspace), clippy clean, fmt clean. Tracker: step 1.5

+++ crates/neunode-inference/src/streaming_settlement.rs
@@ -112,7 +112,7 @@ impl StreamingAccumulator {
             model_info.output_price_per_million,
         );
 
-        let fee_amount = (gross_cost.0 * config.protocol_fee_bps).div_ceil(10_000);
+        let fee_amount = (gross_cost.0 * config.protocol_fee_bps as u128).div_ceil(10_000);
         let protocol_fee = TokenAmount(fee_amount);
         let net_payout = gross_cost
             .checked_sub(protocol_fee)
